bug: correção erros monitor
