+++ src/components/MyCard/index.jsx
@@ -1,0 +1,5 @@
+import { Card } from "@mui/material";
+
+export default function index() {
+  return <div>index</div>;
+}
