+++ src/components/index.tsx
@@ -4,3 +4,12 @@ export { default as DatePickerComponent } from "./datePicker";
 export { default as DateTimePickerComponent } from "./dateTimePicker";
 export { default as SwitchComponent } from "./Switch";
 export { default as TabComponent } from "./Tab";
+export { default as Alert } from "./Alert";
+export { default as MyAvatar } from "./MyAvatar";
+export { default as MyBox } from "./MyBox";
+export { default as MyCard } from "./MyCard";
+export { default as MyCheckBox } from "./MyCheckBox";
+export { default as MyContainer } from "./MyContainer";
+export { default as MySnackBar } from "./MySnackBar";
+export { default as MyFab } from "./MyFab";
+
